Feature/CMakeAndMathFix (#128)
* [Changed] merged all math / number related headers into one util header

* [Fix] wring include

* Delete AnimatedRifle.glb

* [Fix] missing harfbuzz

* [Fix] RIGHT direction [Fix] missing namespaces in Vec3Range
