--- assets/models/AnimatedRifle.glb
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:19533629c0a19b5e35e86c8e20410a193958580f080a869b9770d698b0c116b4
-size 15934896
