--- engine/utility/public/range.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-#include <cstdint>
-
-struct Range
-{
-    uint32_t start;
-    uint32_t count;
-};
